Make the following changes:
,

Find this element:

  <div class="p-6 pt-0">

with the text content:

    Track

Relevant files:

And change it as follows:

I want the bus tracking feature to filter the buses based on their scheduled time. For example, I want to track the bus that departs on 8am from Mantaongon I would choose the time of that bus.