Make the following changes:
,

Find this element:

  <Button className="w-full bg-primary hover:bg-primary/90 text-primary-foreground text-lg py-6" disabled=false>

with the text content:

    Confirm Reservation

Relevant files:

    - https://6000-firebase-studio-1747881705720.cluster-xpmcxs2fjnhg6xvn446ubtgpio.cloudworkstations.dev/_next/static/chunks/src_aa19bdf7._.js
    - https://6000-firebase-studio-1747881705720.cluster-xpmcxs2fjnhg6xvn446ubtgpio.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747881705720.cluster-xpmcxs2fjnhg6xvn446ubtgpio.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

when I click confirm, there's a logic that I still need to Signin or login to the app in order to actually reserve it and if the user is already logged-in, then it will redirect to the payment page, where they need to pay 30% of the fare that they would pay to reserve it. and pay the rest to the bus conductor after the ride. They can also pay the whole fare directly.